Move ...Client/Service classes from OrbitCore to OrbitGl/OrbitService
