+++ OrbitGl/ProcessMemoryClient.h
@@ -5,6 +5,7 @@
 #ifndef ORBIT_CORE_PROCESS_MEMORY_CLIENT_H_
 #define ORBIT_CORE_PROCESS_MEMORY_CLIENT_H_
 
+#include "CoreApp.h"
 #include "TransactionClient.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/synchronization/mutex.h"
@@ -19,18 +20,15 @@ class ProcessMemoryClient {
   ProcessMemoryClient(ProcessMemoryClient&&) = delete;
   ProcessMemoryClient& operator=(ProcessMemoryClient&&) = delete;
 
-  using ProcessMemoryCallback =
-      std::function<void(const std::vector<uint8_t>&)>;
-
   void GetRemoteMemory(uint32_t pid, uint64_t address, uint64_t size,
-                       const ProcessMemoryCallback& callback);
+                       const CoreApp::ProcessMemoryCallback& callback);
 
  private:
   void HandleResponse(const Message& message, uint64_t id);
 
   TransactionClient* transaction_client_;
   absl::Mutex transaction_mutex_;
-  absl::flat_hash_map<uint64_t, ProcessMemoryCallback> callbacks_;
+  absl::flat_hash_map<uint64_t, CoreApp::ProcessMemoryCallback> callbacks_;
 };
 
 #endif  // ORBIT_CORE_PROCESS_MEMORY_CLIENT_H_
