+++ OrbitService/ProcessMemoryService.cpp
@@ -4,14 +4,6 @@
 
 #include "ProcessMemoryService.h"
 
-#include "Serialization.h"
-
-ORBIT_SERIALIZE(ProcessMemoryRequest, 0) {
-  ORBIT_NVP_VAL(0, pid);
-  ORBIT_NVP_VAL(0, address);
-  ORBIT_NVP_VAL(0, size);
-}
-
 ProcessMemoryService::ProcessMemoryService(
     TransactionService* transaction_service)
     : transaction_service_{transaction_service} {
