+++ OrbitService/ProcessMemoryService.h
@@ -7,20 +7,8 @@
 
 #include <cstdint>
 
-#include "SerializationMacros.h"
+#include "ProcessMemoryRequest.h"
 #include "TransactionService.h"
-
-struct ProcessMemoryRequest {
-  ProcessMemoryRequest() = default;
-  ProcessMemoryRequest(uint32_t pid, uint64_t address, uint64_t size)
-      : pid{pid}, address{address}, size{size} {}
-
-  uint32_t pid = 0;
-  uint64_t address = 0;
-  uint64_t size = 0;
-
-  ORBIT_SERIALIZABLE;
-};
 
 class ProcessMemoryService {
  public:
